fix swift lint
